+++ finrl/meta/env_stock_trading/env_stocktrading.py
@@ -13,8 +13,6 @@ from stable_baselines3.common.vec_env import DummyVecEnv
 
 matplotlib.use("Agg")
 
-# from stable_baselines3.common.logger import Logger, KVWriter, CSVOutputFormat
-
 
 class StockTradingEnv(gym.Env):
     """