+++ finrl/train.py
@@ -125,25 +125,7 @@ if __name__ == "__main__":
 
     ## if users want to use rllib, or stable-baselines3, users can remove the following comments
 
-    # # demo for rllib
-    # import ray
-    # ray.shutdown()  # always shutdown previous session if any
-    # train(
-    #     start_date=TRAIN_START_DATE,
-    #     end_date=TRAIN_END_DATE,
-    #     ticker_list=DOW_30_TICKER,
-    #     data_source="yahoofinance",
-    #     time_interval="1D",
-    #     technical_indicator_list=INDICATORS,
-    #     drl_lib="rllib",
-    #     env=env,
-    #     model_name="ppo",
-    #     cwd="./test_ppo",
-    #     rllib_params=RLlib_PARAMS,
-    #     total_episodes=30,
-    # )
-    #
-    # # demo for stable-baselines3
+    # demo for stable-baselines3
     # train(
     #     start_date=TRAIN_START_DATE,
     #     end_date=TRAIN_END_DATE,